requirements.txt
openai>=1.0.0
python-dotenv
requests
tqdm
Pillow
playwright
